Proper cheatcodes
cargo.lock
