+++ huff_tests/Cargo.toml
@@ -21,6 +21,8 @@ comfy-table = "6.0.0"
 yansi = "0.5.1"
 serde = "1.0.143"
 serde_json = "1.0.83"
+phf = { version = "0.11.1", features = ["macros"] }
+lazy_static = "1.4.0"
 huff_parser = { path = "../huff_parser" }
 huff_core = { path = "../huff_core" }
 huff_codegen = { path = "../huff_codegen" }
